requirements.txt
asgiref==3.6.0
Django==4.1.5
et-xmlfile==1.1.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
